+++ src/components/imports/LoginForm.jsx
@@ -1,13 +1,29 @@
 import React from 'react'
 import ErrorAlter from './ErrorAlter';
+import { useLocation, useHistory, Link } from 'react-router-dom';
 
 export default function LoginForm(props) {
 
     let { handleSubmit, setEmail, setPassword, errorMessage, setError } = props.loginState;
+    const location = useLocation();
+    const history = useHistory();
+
+    let defaultClass = "nav-link link-btn btn-primary default-bg";
+    let active = " active"
+
+    const register = () => history.push("/register")
+    const login = () => history.push("/login")
 
     return (
         <div className="col-sm-6  bg-color align-self-center">
- 
+        <div className="info">
+            <div className="btn-section">
+                <button onClick={register} className={location.pathname === "/register" ? defaultClass + active : defaultClass}>Register</button>
+            </div>
+            <div className="btn-section-l">
+                <button onClick={login} className={location.pathname === "/login" ? defaultClass + active : defaultClass}>Login</button>
+            </div>
+        </div>
         <div className="form-section">
             <div className="title-l">
             <h1>DESIGN. DEVELOP. DEPLOY.</h1>
