+++ src/components/imports/RegisterForm.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import ErrorAlert from './ErrorAlter';
+import { useLocation, useHistory, Link } from 'react-router-dom';
 
 export default function RegisterForm(props) {
    
@@ -12,9 +13,27 @@ export default function RegisterForm(props) {
         errorMessage, 
         setError 
         } = props.registerState;
+
+    const location = useLocation();
+    const history = useHistory();
+    
+    let defaultClass = "nav-link link-btn btn-primary default-bg";
+    let active = " active"
+    
+    const register = () => history.push("/register")
+    const login = () => history.push("/login")
+    
    
     return (
         <div className="col-sm-6 bg-color align-self-center">
+            <div className="info-r">
+                <div className="btn-section">
+                    <button onClick={register} className={location.pathname === "/register" ? defaultClass + active : defaultClass}>Register</button>
+                </div>
+                <div className="btn-section-l">
+                    <button onClick={login} className={location.pathname === "/login" ? defaultClass + active : defaultClass}>Login</button>
+                </div>
+            </div>
             <div className="form-section">
                     <div className="title-r">
                         <h1>DESIGN. DEVELOP. DEPLOY.</h1>
